+++ src/tokenomics/contracts/farmings/eternalFarming/interfaces/IAlgebraEternalVirtualPool.sol
@@ -10,10 +10,10 @@ interface IAlgebraEternalVirtualPool is IAlgebraVirtualPoolBase {
     /// @param rate1 The new rate of bonus token distribution per sec
     function setRates(uint128 rate0, uint128 rate1) external;
 
-    function addRewards(uint256 token0Amount, uint256 token1Amount) external;
+    function addRewards(uint128 token0Amount, uint128 token1Amount) external;
 
-    function getInnerRewardsGrowth(int24 bottomTick, int24 topTick)
-        external
-        view
-        returns (uint256 rewardGrowthInside0, uint256 rewardGrowthInside1);
+    function getInnerRewardsGrowth(
+        int24 bottomTick,
+        int24 topTick
+    ) external view returns (uint256 rewardGrowthInside0, uint256 rewardGrowthInside1);
 }
